Fix: invoice pagination, sorting
